[skiplang] Split type definition from wasm runtime
