+++ package.json
@@ -2,6 +2,7 @@
   "workspaces": [
     "eslint-config",
     "tsconfig",
+    "skiplang/prelude/ts/binding",
     "skiplang/prelude/ts/wasm",
     "skiplang/skdate/ts",
     "skiplang/skjson/ts",
