+++ skiplang/prelude/ts/binding/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@skiplang/std",
+  "version": "0.0.1",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js",
+    "./internal.js": "./dist/internal.js"
+  },
+  "scripts": {
+    "build": "tsc",
+    "clean": "rm -rf dist",
+    "lint": "eslint"
+  }
+}
