+++ skiplang/prelude/ts/binding/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@skiplabs/tsconfig"
+}
