+++ skiplang/prelude/ts/wasm/package.json
@@ -4,7 +4,6 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.js",
-    "./internal.js": "./dist/sk_internal_types.js",
     "./runtime.js": "./dist/sk_runtime.js",
     "./posix.js": "./dist/sk_posix.js",
     "./worker.js": "./dist/sk_worker.js"
@@ -18,6 +17,7 @@
     "@types/ws": "^8.5.13"
   },
   "dependencies": {
+    "@skiplang/std": "0.0.1",
     "ws": "^8.18.0"
   }
 }
