+++ skiplang/prelude/ts/wasm/src/sk_posix.ts
@@ -1,5 +1,5 @@
+import type { int } from "@skiplang/std";
 import type {
-  int,
   ptr,
   Environment,
   Links,
@@ -8,7 +8,7 @@ import type {
   FileSystem,
 } from "./sk_types.js";
 import { Options } from "./sk_types.js";
-import type * as Internal from "./sk_internal_types.js";
+import type * as Internal from "@skiplang/std/internal.js";
 
 interface ToWasm {
   SKIP_check_if_file_exists: (skPath: ptr<Internal.String>) => boolean;
